bugfixes, added nested comment.author support

+++ backend/migrations/20210726171949-create-comment.js
@@ -12,7 +12,7 @@ module.exports = {
         type: Sequelize.SMALLINT
       },
       stars: {
-        type: Sequelize.SMALLINT
+        type: Sequelize.FLOAT
       },
       content: {
         type: Sequelize.STRING
